chore: reorganise stuff
