+++ src/OweMe.Api/Identity/Configuration/BearerSecuritySchemeTransformer.cs
@@ -2,7 +2,7 @@
 using Microsoft.AspNetCore.OpenApi;
 using Microsoft.OpenApi.Models;
 
-namespace OweMe.Api.Configuration.Identity;
+namespace OweMe.Api.Identity.Configuration;
 
 internal sealed class BearerSecuritySchemeTransformer(IAuthenticationSchemeProvider authenticationSchemeProvider) : IOpenApiDocumentTransformer
 {
